metaverse rpc with aura consensus
