+++ Makefile
@@ -4,7 +4,7 @@ init:
 
 .PHONY: check
 check: githooks
-	SKIP_WASM_BUILD= cargo check
+	SKIP_WASM_BUILD= cargo check --features with-metaverse-runtime
 
 .PHONY: check-debug
 check-debug:
